+++ app/src/main/res/layout/review_recycler_item_recent.xml
@@ -14,7 +14,7 @@
         android:layout_width="0dp"
         android:layout_height="150dp"
         android:layout_weight="1"
-        android:src="@drawable/user_sample"/>
+        android:src="@drawable/sample_woman"/>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="0dp"
